refactor: type-level pin api invariants for response bodies
