+++ src/body/mod.rs
@@ -3,10 +3,11 @@ mod request;
 mod response;
 mod size_hint;
 
-pub(crate) use either::Either;
-
 pub use bytes::Bytes;
 pub use hyper::body::Frame;
 
 pub use request::{BodyStream, ReadIntoBytes, ReadIntoString, RequestBody};
 pub use response::ResponseBody;
+
+pub(crate) use either::Either;
+pub(crate) use response::Pollable;
