+++ src/body/response/mod.rs
@@ -1,9 +1,11 @@
 mod body;
 mod buffered;
 mod mapped;
+mod pollable;
 mod streaming;
 
 pub use body::ResponseBody;
+pub use pollable::Pollable;
 
 use super::Either;
 use buffered::Buffered;
